src/requirements.txt
Flask==2.2.2
PyNaCl==1.5.0
requests==2.28.1
python-dotenv==0.21.0